Add supprot for encoding numeric types as string
