+++ packages/openapi3/src/schema-emitter.ts
@@ -73,8 +73,10 @@ import {
   shouldInline,
 } from "@typespec/openapi";
 import { getOneOf, getRef } from "./decorators.js";
+import { applyEncoding } from "./encoding.js";
 import { OpenAPI3EmitterOptions, reportDiagnostic } from "./lib.js";
 import { ResolvedOpenAPI3EmitterOptions } from "./openapi.js";
+import { getSchemaForStdScalars } from "./std-scalar-schemas.js";
 import { OpenAPI3Discriminator, OpenAPI3Schema, OpenAPI3SchemaProperty } from "./types.js";
 import { VisibilityUsageTracker } from "./visibility-usage.js";
 
@@ -718,66 +720,7 @@ export class OpenAPI3SchemaEmitter extends TypeEmitter<
   }
 
   #getSchemaForStdScalars(scalar: Scalar & { name: IntrinsicScalarName }): OpenAPI3Schema {
-    switch (scalar.name) {
-      case "bytes":
-        return { type: "string", format: "byte" };
-      case "numeric":
-        return { type: "number" };
-      case "integer":
-        return { type: "integer" };
-      case "int8":
-        return { type: "integer", format: "int8" };
-      case "int16":
-        return { type: "integer", format: "int16" };
-      case "int32":
-        return { type: "integer", format: "int32" };
-      case "int64":
-        return { type: "integer", format: "int64" };
-      case "safeint":
-        switch (this.#options.safeintStrategy) {
-          case "double-int":
-            return { type: "integer", format: "double-int" };
-          case "int64":
-          default:
-            return { type: "integer", format: "int64" };
-        }
-      case "uint8":
-        return { type: "integer", format: "uint8" };
-      case "uint16":
-        return { type: "integer", format: "uint16" };
-      case "uint32":
-        return { type: "integer", format: "uint32" };
-      case "uint64":
-        return { type: "integer", format: "uint64" };
-      case "float":
-        return { type: "number" };
-      case "float64":
-        return { type: "number", format: "double" };
-      case "float32":
-        return { type: "number", format: "float" };
-      case "decimal":
-        return { type: "number", format: "decimal" };
-      case "decimal128":
-        return { type: "number", format: "decimal128" };
-      case "string":
-        return { type: "string" };
-      case "boolean":
-        return { type: "boolean" };
-      case "plainDate":
-        return { type: "string", format: "date" };
-      case "utcDateTime":
-      case "offsetDateTime":
-        return { type: "string", format: "date-time" };
-      case "plainTime":
-        return { type: "string", format: "time" };
-      case "duration":
-        return { type: "string", format: "duration" };
-      case "url":
-        return { type: "string", format: "uri" };
-      default:
-        const _assertNever: never = scalar.name;
-        return {};
-    }
+    return getSchemaForStdScalars(scalar, this.#options);
   }
 
   #applySchemaExamples(
@@ -904,31 +847,16 @@ export class OpenAPI3SchemaEmitter extends TypeEmitter<
     typespecType: Scalar | ModelProperty,
     target: OpenAPI3Schema | Placeholder<OpenAPI3Schema>
   ): OpenAPI3Schema {
-    const encodeData = getEncode(this.emitter.getProgram(), typespecType);
-    if (encodeData) {
-      const newTarget = new ObjectBuilder(target);
-      const newType = this.#getSchemaForStdScalars(encodeData.type as any);
-
-      newTarget.type = newType.type;
-      // If the target already has a format it takes priority. (e.g. int32)
-      newTarget.format = this.#mergeFormatAndEncoding(
-        newTarget.format,
-        encodeData.encoding,
-        newType.format
-      );
-      return newTarget;
-    }
-    const result = new ObjectBuilder(target);
-    return result;
+    return applyEncoding(this.emitter.getProgram(), typespecType, target as any, this.#options);
   }
   #mergeFormatAndEncoding(
     format: string | undefined,
-    encoding: string,
+    encoding: string | undefined,
     encodeAsFormat: string | undefined
-  ): string {
+  ): string | undefined {
     switch (format) {
       case undefined:
-        return encodeAsFormat ?? encoding;
+        return encodeAsFormat ?? encoding ?? format;
       case "date-time":
         switch (encoding) {
           case "rfc3339":
